requirement.txt
altair==5.3.0
appnope==0.1.4
asttokens==2.4.1
attrs==23.2.0
backcall==0.2.0
bcrypt==4.1.3
blinker==1.8.2
cachetools==5.3.3
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
comm==0.2.2
contourpy==1.1.1
cycler==0.12.1
debugpy==1.8.1
decorator==5.1.1
executing==2.0.1
extra-streamlit-components==0.1.71
ezodf==0.3.2
fonttools==4.51.0
gitdb==4.0.11
GitPython==3.1.43
idna==3.7
importlib_metadata==7.1.0
importlib_resources==6.4.0
ipykernel==6.29.4
ipython==8.12.3
jedi==0.19.1
Jinja2==3.1.4
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
jupyter_client==8.6.1
jupyter_core==5.7.2
kiwisolver==1.4.5
lxml==4.9.4
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.7.5
matplotlib-inline==0.1.7
mdurl==0.1.2
nest-asyncio==1.6.0
numpy==1.24.4
packaging==24.0
pandas==1.5.3
pandas-ods-reader==0.1.4
parso==0.8.4
pexpect==4.9.0
pickleshare==0.7.5
pillow==10.3.0
pkgutil_resolve_name==1.3.10
platformdirs==4.2.2
prompt-toolkit==3.0.43
protobuf==4.25.3
psutil==5.9.8
psycopg-binary==3.1.18
psycopg2==2.9.9
psycopg2-binary==2.9.9
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==16.1.0
pydeck==0.9.1
Pygments==2.18.0
PyJWT==2.8.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
pyzmq==26.0.3
referencing==0.35.1
requests==2.31.0
rich==13.7.1
rpds-py==0.18.1
six==1.16.0
smmap==5.0.1
SQLAlchemy==2.0.30
stack-data==0.6.3
streamlit==1.34.0
streamlit-authenticator==0.3.2
streamlit-card==1.0.2
streamlit-option-menu==0.3.12
tenacity==8.3.0
toml==0.10.2
toolz==0.12.1
tornado==6.4
traitlets==5.14.3
typing_extensions==4.11.0
tzdata==2024.1
urllib3==2.2.1
watchdog==4.0.0
wcwidth==0.2.13
zipp==3.18.1
